Adding CI Pipeline to the Devel Branch, initial test (#21)
Done:
* Moved files to WormJam Core repo from test repository
* Added Travis CI/Discord integration
* Fixed minor text parsing in tsv files from initial creation
* Changed the structure of the tests to give user better feedback
* Update to github-pages
* Updated hardcoded locations of files in processing scripts
* Cleaned up duplicate python scripts

Still to do:
* Rework gh-pages
* Improve the parsing of TSVs to SBML format.
