+++ travis/basic_fba.py
@@ -1,6 +1,6 @@
 import cobra
 
-model = cobra.io.read_sbml_model("test_model.xml")
+model = cobra.io.read_sbml_model("WormJam.xml")
 print("Model:")
 print(len(model.reactions),"reactions")
 print(len(model.metabolites),"metabolites")
@@ -9,8 +9,8 @@ print(len(model.genes),"genes")
 biomass_rxn = model.reactions.get_by_id("BIO0100")
 model.objective = biomass_rxn
 medium = model.medium
-for i in medium:
-    medium[i] = 0
+medium["O2_Exchange_reactions_e"] = 1000
+medium["EXC0050"] = 1000
 model.medium = medium
 with model:
     solution = model.optimize()
@@ -19,4 +19,4 @@ print("Solution:")
 print(solution.objective_value)
 print(solution.status)
 
-assert solution.objective_value== 0, "Flux carried under restricted conditions"
+assert solution.objective_value!= 0, "Flux not carried in normal growth"
